pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
